+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Breaking changes introduced. Bumped Testify to 1.0.0-beta1
 - Correctly access external files directory on API 28+
 - Correctly detect the locale on API 22
 - `SoftwareRenderViewModification` now qualifies all views, not just `ImageView`
+- setFontScale now works correctly on API 21+
 
 #### Testing changes